Loading first lawn in list.
The lawn component now gets a list of lawns in database when mounting and adds them as select options. When selecting a lawn, a database call is made and assigns the component the data. The columns are set in the data and the rows are computed from the number of tiles and the columns.